badsigs.sh is just running badsigs.Main Scala.
It's worth noting that badsigs depends on existense
of a jar produced by `sbt assembly` command.

+++ badsigs.sh
@@ -11,6 +11,13 @@ then
   exit 1
 fi
 
+BADSIGS_JAR="target/badsigs-assembly-0.1-SNAPSHOT.jar"
+if [ ! -f "$BADSIGS_JAR" ]
+then
+  echo "Compile project with sbt 0.10.x, first. Run \`assembly\` command."
+  exit 1
+fi
+
 CD=`pwd`
 WD="$CD/badsigs_working_dir"
 [ -d $WD ] && rm -rf $WD
@@ -33,17 +40,14 @@ echo "Checking $SCALA_JAR:" | tee -a $LOG
 cat $TMP/library.properties | tee -a $LOG
 echo "" | tee -a $LOG
 
-echo "Generating Java files"
-scala -classpath target/scala-2.9.0.1/classes/ badsigs.Main $TMP $SRC
-N=`find $SRC -name '*.java' | wc -l`
-echo "Generated $N files"
+echo "Running Main app (will generate Java files and run ecj)"
+java -jar "$BADSIGS_JAR" $TMP $SRC 2>> $LOG
 
-echo "Running ecj"
-for i in `find $SRC -name '*.java'`; do
-  echo "Processing $i"
-  java -jar ecj-3.7.jar -1.5 -nowarn -d none -classpath $SCALA_JAR $i 2>> $LOG
-  [ $? -ne 0 ] && echo "" >> $LOG
-done
-echo ""
-
-echo "Check $LOG for errors."
+if [ $? -eq 0 ]
+then
+  echo "No errors found!"
+  exit 0
+else
+  echo "Errors found. Check $LOG for details."
+  exit 1
+fi
